Create GithubRepoSearchAPIRequest.swift, GithubAPIError.swift
